Struct tres::Traced

pub struct Traced<E, T = Locations> where
    T: Trace,  { /* private fields */ }

Wraps a generic error value and keeps track of an error trace.

Implementations

impl<E, T> Traced<E, T> where
T: Trace + Default,

pub fn new(inner: E) -> Self

Wraps the given error and starts a new trace with the caller’s location.

Examples

Basic usage:

use tres::Traced;

let x: Traced<&str> = Traced::new("Oops!");

Showing that the trace includes the caller of new():

use std::panic::Location;
use tres::{Locations, Traced};

let here: &Location = Location::caller();
let x: Traced<&str> = Traced::new("Oops!");

let locs: &Locations = x.trace();
let there: &Location = locs.0.first().unwrap();
assert_eq!(there.line(), here.line() + 1);

Using a custom trace type:

use std::panic::Location;
use tres::{Trace, Traced};

#[derive(Default)]
struct BangTrace(pub String);

impl Trace for BangTrace {
    fn trace(&mut self, _location: &'static Location) {
        self.0.push('!');
    }
}

let x: Traced<&str, BangTrace> = Traced::new("Oops!");
assert_eq!(&x.trace().0, "!");

pub fn empty(inner: E) -> Self

Wraps the given error and starts a new, empty trace.

Examples

Basic usage:

use tres::Traced;

let x: Traced<&str> = Traced::empty("Oops!");
assert!(x.trace().0.is_empty());

impl<E, T> Traced<E, T> where
T: Trace,

pub fn inner(&self) -> &E

Returns a reference to the contained error value.

Examples

Basic usage:

use tres::Traced;

let x: Traced<String> = Traced::new("Oops!".to_string());

let inner: &String = x.inner();
assert_eq!(inner.as_str(), "Oops!");

pub fn trace(&self) -> &T

Returns a reference to the error trace.

Examples

Basic usage:

use tres::{Locations, Traced};

let x: Traced<String> = Traced::new("Oops!".to_string());

let trace: &Locations = x.trace();
assert_eq!(trace.0.len(), 1);

pub fn from_parts(inner: E, trace: T) -> Self

Constructs a new Traced from an error value and a trace.

Examples

Basic usage:

use tres::{Locations, Traced};

let error: String = "Oops!".into();
let trace: Locations = Default::default();

let x = Traced::from_parts(error, trace);
assert_eq!(x.inner(), &"Oops!");
assert_eq!(x.trace(), &Locations(vec![]));

pub fn into_parts(self) -> (E, T)

Returns the contained error value and error trace, consuming self.

Examples

Basic usage:

use tres::{Locations, Traced};

let x: Traced<String> = Traced::new("Oops!".to_string());

let (error, trace): (String, Locations) = x.into_parts();
assert_eq!(error, "Oops!".to_string());
assert_eq!(trace.0.len(), 1);

pub fn into_inner(self) -> E

Returns the contained error value, consuming self and discarding the trace.

Examples

Basic usage:

use tres::Traced;

let x: Traced<String> = Traced::new("Oops!".to_string());

let error: String = x.into_inner();
assert_eq!(error, "Oops!".to_string());

pub fn map<F, O: FnOnce(E) -> F>(self, op: O) -> Traced<F, T>

Maps a Traced<E, T> to Traced<F, T> by applying a function to the contained error value, leaving the error trace untouched.

Examples

Basic usage:

use tres::Traced;

let x: Traced<u32> = Traced::new(42);
assert_eq!(x.trace().0.len(), 1);

let x: Traced<String> = x.map(|i| i.to_string());
assert_eq!(x.trace().0.len(), 1);